+++ packages/multichain-account-service/jest.config.js
@@ -20,7 +20,7 @@ module.exports = merge(baseConfig, {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 100,
+      branches: 96.76,
       functions: 100,
       lines: 100,
       statements: 100,